Move CI info with the other execution info
