+++ react/javascript/src/components/app/FilteredResults.tsx
@@ -13,7 +13,6 @@ import StatusesFilter from './StatusesFilter'
 import countScenariosByStatuses from '../../countScenariosByStatuses'
 import ExecutionSummary from './ExecutionSummary'
 import EnvelopesQueryContext from '../../EnvelopesQueryContext'
-import CIInformation from './CIInformation'
 
 const defaultDisplayedResults = [
   messages.TestStepFinished.TestStepResult.Status.AMBIGUOUS,
@@ -67,8 +66,6 @@ const FilteredResults: React.FunctionComponent = () => {
           scenarioCountByStatus={scenarioCountByStatus}
         />
         <ExecutionSummary meta={meta} />
-        {meta.ci && <CIInformation ci={meta.ci} />}
-
         <SearchBar queryUpdated={(query) => setQuery(query)} />
       </div>
       <GherkinDocumentList gherkinDocuments={filtered} />
